requirements.txt
flask==2.3.3
openai-whisper==20231117
openai>=1.0.0
torch==2.1.0
torchaudio==2.1.0
werkzeug==2.3.7
numpy==1.24.3
ffmpeg-python==0.2.0
